stockanalyst_agent/requirements.txt

google-adk
opentelemetry-instrumentation-google-genai
yfinance
yahoo-finance-server
google-adk[eval]
google-adk[A2A]
pydantic
cloudpickle
google-cloud-aiplatform

